+++ src/ts/types/Stage.ts
@@ -12,7 +12,6 @@
 
 import * as sass from 'sass';
 import * as typeDoc from "typedoc";
-import typescript from 'typescript';
 
 import type {
     Json,
@@ -608,6 +607,19 @@ export namespace Stage {
          */
         export interface Release extends Args {
 
+            /**
+             * Files to add to commit after packaging but before release.
+             */
+            commit: null | (
+                /**
+                 * @param stage         Instance of the current stage (probably 
+                 *                      {@link ReleaseStage}).
+                 * @param defaultPaths  The default paths to add to the commit.
+                 * 
+                 * @return  All relative or absolute paths to add to the commit.
+                 */
+                ( stage: Stage, defaultPaths?: string[] ) => string[]
+            );
 
             /**
              * Whether to include this sub-stage, or the configuration if so.
@@ -834,6 +846,7 @@ export namespace Stage {
          * Optional configuration for {@link Compiler} classes.
          * 
          * @since 0.1.0-alpha
+         * @since 0.1.1 — Removed unused ts prop.
          */
         export interface Args {
 
@@ -842,10 +855,10 @@ export namespace Stage {
              */
             sass: sass.Options<"sync">;
 
-            /**
-             * Optional default configuration to use when compiling typescript.
-             */
-            ts: typescript.CompilerOptions,
+            // /**
+            //  * Optional default configuration to use when compiling typescript.
+            //  */
+            // ts: typescript.CompilerOptions,
         };
     };